+++ rpc/core/types/responses.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 	"time"
 
-	abci "github.com/tendermint/abci/types"
+	abci "github.com/tendermint/tendermint/abci/types"
 	crypto "github.com/tendermint/tendermint/crypto"
 	cmn "github.com/tendermint/tmlibs/common"
 
@@ -65,7 +65,7 @@ type SyncInfo struct {
 	LatestAppHash     cmn.HexBytes `json:"latest_app_hash"`
 	LatestBlockHeight int64        `json:"latest_block_height"`
 	LatestBlockTime   time.Time    `json:"latest_block_time"`
-	Syncing           bool         `json:"syncing"`
+	CatchingUp        bool         `json:"catching_up"`
 }
 
 // Info about the node's validator